stock_pricing/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==1.16.0
Flask==3.0.3
frozendict==2.4.4
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
googleapis-common-protos==1.63.2
grpcio==1.65.2
grpcio-status==1.62.2
html5lib==1.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
peewee==3.17.6
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
soupsieve==2.5
statsmodels==0.14.2
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.40
zipp==3.19.2
